deckgen/requirements.txt
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
scipy==1.11.4
ttkbootstrap==1.10.1